Initial code structure for the lavocado
The lavocado aims to be an alternative test framework for the libvirt
project using as muscles only Avocado and python-libvirt. This can be
used to write unit, functional and integration tests and it is inspired
on the libvirt-tck framework, but instead of Perl it written in Python.

The idea is to provide to the test writers helper classes to avoid
boilerplate code, improve readability and code maintenance.

Signed-off-by: Beraldo Leal <bleal@redhat.com>